// backend/src/types.ts
export enum Role {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  OPERATIONS_MANAGER = 'OPERATIONS_MANAGER',
  DOCUMENTATION_EXECUTIVE = 'DOCUMENTATION_EXECUTIVE',
  CUSTOMS_EXECUTIVE = 'CUSTOMS_EXECUTIVE',
  TRANSPORT_MANAGER = 'TRANSPORT_MANAGER',
  SALES_EXECUTIVE = 'SALES_EXECUTIVE',
  CUSTOMER = 'CUSTOMER',
  ACCOUNTANT = 'ACCOUNTANT',
  EMPLOYEE = 'EMPLOYEE'
}